convert to package format
